refactoring after code review
