+++ src/features/index.ts
@@ -1,4 +1,5 @@
 export * from './auth'
+export * from './notifications'
 export * from './posts'
 export * from './profile'
 export * from './public'
